remove redundant evaluation code (#315)
