+++ deeplite_torch_zoo/api/eval/yolo.py
@@ -1,4 +1,66 @@
-from deeplite_torch_zoo.src.object_detection.eval.yolov5_eval.wrappers import yolo_eval_coco  # pylint: disable=unused-import
-from deeplite_torch_zoo.src.object_detection.eval.yolov5_eval.wrappers import yolo_eval_voc  # pylint: disable=unused-import
-from deeplite_torch_zoo.src.object_detection.eval.zoo_eval.extra_eval.lisa_eval import yolo_eval_lisa  # pylint: disable=unused-import
-from deeplite_torch_zoo.src.object_detection.eval.zoo_eval.extra_eval.wider_face_eval import yolo_eval_wider_face  # pylint: disable=unused-import
+import numpy as np
+
+from deeplite_torch_zoo.src.object_detection.eval.evaluate import evaluate
+from deeplite_torch_zoo.api.registries import EVAL_WRAPPER_REGISTRY
+
+
+@EVAL_WRAPPER_REGISTRY.register(
+    task_type='object_detection', model_type='yolo', dataset_type='voc07'
+)
+@EVAL_WRAPPER_REGISTRY.register(
+    task_type='object_detection', model_type='yolo', dataset_type='voc'
+)
+def yolo_eval_voc(
+    model,
+    test_dataloader,
+    device="cuda",
+    iou_thresh=0.5,
+    conf_thresh=0.001,
+    nms_thresh=0.5,
+    eval_style='coco',
+    progressbar=False,
+    subclasses=None,
+    num_classes=None,
+    v8_eval=False,
+    **kwargs
+):
+    model.to(device)
+    ap_dict = evaluate(
+        model,
+        test_dataloader,
+        conf_thres=conf_thresh,
+        iou_thres=nms_thresh,
+        eval_style=eval_style,
+        map_iou_thresh=iou_thresh,
+        num_classes=num_classes,
+        v8_eval=v8_eval,
+    )
+    return ap_dict
+
+
+@EVAL_WRAPPER_REGISTRY.register(
+    task_type='object_detection', model_type='yolo', dataset_type='coco'
+)
+def yolo_eval_coco(
+    model,
+    test_dataloader,
+    device="cuda",
+    conf_thresh=0.001,
+    nms_thresh=0.5,
+    eval_style='coco',
+    progressbar=False,
+    subclasses=None,
+    num_classes=None,
+    **kwargs
+):
+    model.to(device)
+    ap_dict = evaluate(
+        model,
+        test_dataloader,
+        conf_thres=conf_thresh,
+        iou_thres=nms_thresh,
+        eval_style=eval_style,
+        map_iou_thresh=np.arange(0.5, 1.0, 0.05),
+        num_classes=num_classes,
+    )
+    return ap_dict
