--- deeplite_torch_zoo/src/object_detection/eval/zoo_eval/extra_eval/__init__.py
@@ -1,2 +1,0 @@
-from .lisa_eval import *  # pylint: disable=unused-import
-from .wider_face_eval import *  # pylint: disable=unused-import
